+++ src/packages/piral-hyperapp/src/create.ts
@@ -8,8 +8,12 @@ import { PiralHyperappApi } from './types';
  */
 export function createHyperappApi<T extends PiralCoreApi<any>>(api: T): PiralHyperappApi<T> {
   return {
-    registerTileHyperapp(id, root, state, actions, options) {
-      api.registerTileX(id, (el, props, ctx) => mount(el, root, props, ctx, state, actions), options);
+    registerTileHyperapp(id, root, state, actions, options?) {
+      if (typeof id === 'string') {
+        api.registerTileX(id, (el, props, ctx) => mount(el, root, props, ctx, state, actions), options);
+      } else {
+        api.registerTileX((el, props, ctx) => mount(el, id, props, ctx, root, state), actions);
+      }
     },
     registerPageHyperapp(route, root, state, actions) {
       api.registerPageX(route, (el, props, ctx) => mount(el, root, props, ctx, state, actions));
@@ -17,8 +21,12 @@ export function createHyperappApi<T extends PiralCoreApi<any>>(api: T): PiralHyp
     registerExtensionHyperapp(slot, root, state, actions, defaults) {
       api.registerExtensionX(slot, (el, props, ctx) => mount(el, root, props as any, ctx, state, actions), defaults);
     },
-    registerMenuHyperapp(id, root, state, actions, settings) {
-      api.registerMenuX(id, (el, props, ctx) => mount(el, root, props, ctx, state, actions), settings);
+    registerMenuHyperapp(id, root, state, actions, settings?) {
+      if (typeof id === 'string') {
+        api.registerMenuX(id, (el, props, ctx) => mount(el, root, props, ctx, state, actions), settings);
+      } else {
+        api.registerMenuX((el, props, ctx) => mount(el, id, props, ctx, root, state), actions);
+      }
     },
     registerModalHyperapp(id, root, state, actions, defaults) {
       api.registerModalX(id, (el, props, ctx) => mount(el, root, props as any, ctx, state, actions), defaults);