+++ src/packages/piral-vue/src/create.ts
@@ -8,8 +8,12 @@ import { PiralVueApi } from './types';
  */
 export function createVueApi<T extends PiralCoreApi<any>>(api: T): PiralVueApi<T> {
   return {
-    registerTileVue(id, root, options) {
-      api.registerTileX(id, (el, props, ctx) => mount(el, root, props, ctx), options);
+    registerTileVue(id, root, options?) {
+      if (typeof id === 'string') {
+        api.registerTileX(id, (el, props, ctx) => mount(el, root, props, ctx), options);
+      } else {
+        api.registerTileX((el, props, ctx) => mount(el, id, props, ctx), root);
+      }
     },
     registerPageVue(route, root) {
       api.registerPageX(route, (el, props, ctx) => mount(el, root, props, ctx));
@@ -17,8 +21,12 @@ export function createVueApi<T extends PiralCoreApi<any>>(api: T): PiralVueApi<T
     registerExtensionVue<TOpt>(slot, root, defaults) {
       api.registerExtensionX<TOpt>(slot, (el, props, ctx) => mount(el, root, props, ctx), defaults);
     },
-    registerMenuVue(id, root, settings) {
-      api.registerMenuX(id, (el, props, ctx) => mount(el, root, props, ctx), settings);
+    registerMenuVue(id, root, settings?) {
+      if (typeof id === 'string') {
+        api.registerMenuX(id, (el, props, ctx) => mount(el, root, props, ctx), settings);
+      } else {
+        api.registerMenuX((el, props, ctx) => mount(el, id, props, ctx), root);
+      }
     },
     registerModalVue<TOpt>(id, root, defaults) {
       api.registerModalX<TOpt>(id, (el, props, ctx) => mount(el, root, props, ctx), defaults);