+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Distinguish between Piral development and contributing (#42)
 * Started improving the Piral documentation structure (#41)
 * Adapt documentation about static Piral instance (#43)
+* Make first argument of many register methods optional (#29)
 
 ## 0.6.1 (July 26, 2019)